+++ src/components/controls/controls.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useSelector } from "react-redux";
 import { useTranslation } from 'react-i18next';
 
 import ColorBy, {ColorByInfo} from "./color-by";
@@ -16,84 +17,114 @@ import GeoResolution from "./geo-resolution";
 import TransmissionLines from './transmission-lines';
 import NormalizeFrequencies from "./frequency-normalization";
 import AnimationOptions from "./animation-options";
-import PanelToggles from "./panel-toggles";
+import { PanelSection } from "./panelSection";
 import ToggleTangle from "./toggle-tangle";
 import Language from "./language";
 import { ControlsContainer } from "./styles";
 import FilterData, {FilterInfo} from "./filter";
-import {TreeOptionsInfo, MapOptionsInfo, AnimationOptionsInfo, PanelOptionsInfo,
-  ExplodeTreeInfo, FrequencyInfo, MeasurementsOptionsInfo} from "./miscInfoText";
-import { AnnotatedHeader } from "./annotatedHeader";
+import {TreeInfo, MapInfo, AnimationOptionsInfo, PanelLayoutInfo,
+  ExplodeTreeInfo, EntropyInfo, FrequencyInfo, MeasurementsInfo} from "./miscInfoText";
+import { ControlHeader } from "./controlHeader";
 import MeasurementsOptions from "./measurementsOptions";
+import { RootState } from "../../store";
 
-type Props = {
-  treeOn: boolean
-  mapOn: boolean
-  frequenciesOn: boolean
-  measurementsOn: boolean
-}
-
-function Controls({ treeOn, mapOn, frequenciesOn, measurementsOn }: Props) {
+function Controls() {
   const { t } = useTranslation();
+
+  const panelsAvailable = useSelector((state: RootState) => state.controls.panelsAvailable);
+  const panelsToDisplay = useSelector((state: RootState) => state.controls.panelsToDisplay);
+  const showTreeToo = useSelector((state: RootState) => state.controls.showTreeToo);
+  const canTogglePanelLayout = useSelector((state: RootState) => state.controls.canTogglePanelLayout);
 
   return (
     <ControlsContainer>
       <ChooseDataset />
 
-      <AnnotatedHeader title={t("sidebar:Date Range")} tooltip={DateRangeInfo}/>
+      <ControlHeader title={t("sidebar:Date Range")} tooltip={DateRangeInfo}/>
       <DateRangeInputs />
       <AnimationControls />
 
-      <AnnotatedHeader title={t("sidebar:Color By")} tooltip={ColorByInfo}/>
+      <ControlHeader title={t("sidebar:Color By")} tooltip={ColorByInfo}/>
       <ColorBy />
 
-      <AnnotatedHeader title={t("sidebar:Filter Data")} tooltip={FilterInfo}/>
-      <FilterData measurementsOn={measurementsOn} />
+      <ControlHeader title={t("sidebar:Filter Data")} tooltip={FilterInfo}/>
+      <FilterData measurementsOn={panelsToDisplay.includes("measurements")} />
 
-      {treeOn &&
-        <span>
-          <AnnotatedHeader title={t("sidebar:Tree Options")} tooltip={TreeOptionsInfo}/>
-          <ChooseLayout />
-          <ChooseMetric />
-          <ChooseBranchLabelling />
-          <ChooseTipLabel />
-          <ChooseSecondTree />
-          <ChooseExplodeAttr tooltip={ExplodeTreeInfo} />
-          <ToggleTangle />
-        </span>
+      <span style={{ paddingTop: "10px" }} />
+
+      {panelsAvailable.includes("tree") &&
+        <PanelSection
+          panel="tree"
+          title={t("sidebar:Tree")}
+          tooltip={TreeInfo}
+          options={<>
+            <ChooseLayout />
+            <ChooseMetric />
+            <ChooseBranchLabelling />
+            <ChooseTipLabel />
+            <ChooseSecondTree />
+            <ChooseExplodeAttr tooltip={ExplodeTreeInfo} />
+            <ToggleTangle />
+          </>}
+        />
       }
 
-      {measurementsOn &&
-        <span style={{ marginTop: "10px" }}>
-          <AnnotatedHeader title={t("sidebar:Measurements Options")} tooltip={MeasurementsOptionsInfo}/>
-          <MeasurementsOptions />
-        </span>
+      {panelsAvailable.includes("measurements") &&
+        <PanelSection
+          panel="measurements"
+          title={t("sidebar:Measurements")}
+          tooltip={MeasurementsInfo}
+          options={<MeasurementsOptions />}
+        />
       }
 
-      {mapOn &&
-        <span style={{ marginTop: "10px" }}>
-          <AnnotatedHeader title={t("sidebar:Map Options")} tooltip={MapOptionsInfo}/>
-          <GeoResolution />
-          <TransmissionLines />
-        </span>
+      {/* Prevent the map from being toggled on when a second tree is visible.
+          It is hidden by logic elsewhere.
+      */}
+      {panelsAvailable.includes("map") && !showTreeToo &&
+        <PanelSection
+          panel="map"
+          title={t("sidebar:Map")}
+          tooltip={MapInfo}
+          options={<>
+            <GeoResolution />
+            <TransmissionLines />
+          </>}
+        />
       }
 
-      {frequenciesOn &&
-        <span style={{ marginTop: "10px" }}>
-          <AnnotatedHeader title={t("sidebar:Frequency Options")} tooltip={FrequencyInfo}/>
-          <NormalizeFrequencies />
-        </span>
+      {panelsAvailable.includes("entropy") &&
+        <PanelSection
+          panel="entropy"
+          title={t("sidebar:Entropy")}
+          tooltip={EntropyInfo}
+        />
+      }
+
+      {panelsAvailable.includes("frequencies") &&
+        <PanelSection
+          panel="frequencies"
+          title={t("sidebar:Frequency")}
+          tooltip={FrequencyInfo}
+          options={<NormalizeFrequencies />}
+        />
       }
 
       <span style={{ marginTop: "10px" }}>
-        <AnnotatedHeader title={t("sidebar:Animation Options")} tooltip={AnimationOptionsInfo}/>
+        <ControlHeader title={t("sidebar:Animation Options")} tooltip={AnimationOptionsInfo}/>
         <AnimationOptions />
       </span>
 
+      {canTogglePanelLayout &&
+        <>
+          <span style={{ paddingTop: "10px" }} />
+          <ControlHeader title={t("sidebar:Display")} tooltip={PanelLayoutInfo} />
+          <PanelLayout />
+        </>
+      }
+
       <span style={{ paddingTop: "10px" }} />
-      <AnnotatedHeader title={t("sidebar:Panel Options")} tooltip={PanelOptionsInfo}/>
-      <PanelLayout />
-      <PanelToggles />
+      <ControlHeader title={t("sidebar:Language")}/>
       <Language />
     </ControlsContainer>
   );