+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+* Redesigned the panel sections in the sidebar. ([#1704](https://github.com/nextstrain/auspice/pull/1704))
+    * Moved panel visibility toggles to the header.
+    * Added the ability to show/hide panel options.
+* customisation: Renamed `sidebarTheme.unselectedBackground` to a more generic name, `sidebarTheme.alternateBackground`, while keeping backwards compatibility. ([#1704](https://github.com/nextstrain/auspice/pull/1704/commits/1ad4376bb3be9787181110d47fe8c5f26a207896))
+
 ## version 2.50.0 - 2023/10/30